remove unneeded test-framework
Entire-Checkpoint: 1169c7a67db5
