+++ crates/test-framework/Cargo.toml
@@ -10,12 +10,8 @@ version.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-app = { path = "../app" }
 arrow = { workspace = true, features = ["prettyprint", "chrono-tz"] }
-arrow-json.workspace = true
-async-channel = "2.3"
 async-trait.workspace = true
-bytes.workspace = true
 chrono.workspace = true
 clap = { workspace = true, features = ["derive"] }
 dashmap = "6.1.0"
@@ -24,14 +20,10 @@ flight_client = { path = "../flight_client" }
 futures.workspace = true
 indicatif = "0.18.0"
 insta.workspace = true
-object_store = { workspace = true, features = ["aws", "http", "azure", "gcp"] }
-octocrab = "0.49.5"
 opentelemetry.workspace = true
 opentelemetry-otlp.workspace = true
 opentelemetry_sdk.workspace = true
 otel-arrow = { path = "../otel-arrow" }
-rand.workspace = true
-rcgen = { version = "0.14", features = ["x509-parser"] }
 regex.workspace = true
 reqwest.workspace = true
 rustls.workspace = true
@@ -43,11 +35,9 @@ spicepod = { path = "../spicepod" }
 sysinfo = "0.37.2"
 telemetry = { path = "../telemetry" }
 tempfile.workspace = true
-time = "0.3"
 tokio.workspace = true
 tokio-util.workspace = true
 tonic.workspace = true
-util = { path = "../util" }
 uuid = { workspace = true, features = ["v4"] }
 
 [target.'cfg(not(windows))'.dependencies]
@@ -56,25 +46,3 @@ nix = { version = "0.30", features = ["signal"] }
 [dev-dependencies]
 rstest = "0.26.1"
 
-[features]
-anonymous_telemetry = []
-aws-secrets-manager = []
-clickhouse = []
-databricks = []
-debezium = []
-delta_lake = []
-dremio = []
-flightsql = []
-ftp = []
-imap = []
-keyring-secret-store = []
-models = []
-mssql = []
-mysql = []
-odbc = []
-postgres = []
-sharepoint = []
-snowflake = []
-spark = []
-sqlite = []
-turso = []
